+++ src/applications/disability-benefits/2346/2346-schema.json
@@ -589,7 +589,8 @@
   },
   "required": [
     "privacyAgreementAccepted",
-    "fullName",
+    "veteranAddress",
+    "tempAddress",
     "address",
     "gender",
     "email",
