+++ src/applications/disability-benefits/2346/components/addressFields/addressViewField.jsx
@@ -1,4 +1,3 @@
-import PropTypes from 'prop-types';
 import React from 'react';
 
 export const AddressViewField = ({ formData }) => {
@@ -12,7 +11,9 @@ export const AddressViewField = ({ formData }) => {
 
   return (
     <p className="blue-bar-block">
-      {street && <br />}
+      {street && street}
+      <br />
+      {street2 && street2}
       {street2 && <br />}
       {city}
       {','} {state} {zipString}
@@ -20,22 +21,4 @@ export const AddressViewField = ({ formData }) => {
       {country}
     </p>
   );
-};
-
-AddressViewField.propTypes = {
-  street: PropTypes.string.isRequired,
-  city: PropTypes.string.isRequired,
-  country: PropTypes.string.isRequired,
-  state: PropTypes.string.isRequired,
-  postalCode: PropTypes.string.isRequired,
-};
-
-AddressViewField.defaultProps = {
-  formData: {},
-  street: '',
-  street2: '',
-  city: '',
-  state: '',
-  country: '',
-  postalCode: '',
 };
