+++ src/applications/disability-benefits/2346/sass/form-2346.scss
@@ -14,6 +14,9 @@
   padding-bottom: 1em;
   margin-bottom: 1em;
 }
+.gray-color {
+  background-color: #f1f1f1;
+}
 
 .mdot-alert-body {
   padding-left: 1rem;
